+++ Assets/Scipts/Weapons/ProjectileArrow.cs
@@ -82,7 +82,7 @@ public class ProjectileArrow : MonoBehaviour
 					_playerUnit.PenetrationProjectile.CurrentPenetration++;
 
 					// Уменьшаем урон с каждым пробитием
-					_playerUnit.Damage.Actual = (int)(_playerUnit.Damage.Max * (1 - _playerUnit.PenetrationProjectile.PenetrationDamageDecrease.Value/100f));
+					_playerUnit.Damage.Actual = _playerUnit.Damage.Max * (1 - _playerUnit.PenetrationProjectile.PenetrationDamageDecrease.Value/100f);
 
 					// Если число пробитий подошло к пределу, то удаляем стрелу
 					if (_playerUnit.PenetrationProjectile.CurrentPenetration == _playerUnit.PenetrationProjectile.MaxPenetrationCount.Value)
